ResourceLoader: Improve WikiModule title info preload on OutputPage
The separate batches were ineffective in practice. The original idea
was this:

* Separate modules for logged-in users into $userBatch.
* For logged-out page views, the $userBatch is empty.
* For logged-in page views, the $siteBatch is the same
  as for logged-out users (shared cache, cache-hit benefit),
  and then only a small $userBatch is added on top.

In reality:

* For logged-out page views, $userBatch is indeed empty. OK.

  We queue "user" and "user.styles", and their composition
  (WikiModule::getPages) is an empty array for anon user context,
  so it adds nothing to the actual batch (e.g. print json_encode
  $byDomain in WikiModule to see this).

* For logged-in page views in local development on plain MW core,
  it works by design.

* For logged-in page views when you have GlobalCssJs installed,
  like in WMF production, then it doesn't work as designed, because
  ext.globalCssJs.user.styles is a styles module that is user-specific
  but ends up in $siteBatch, thus there is no shared cache with
  logged-out because their $siteBatch is different.

Simplify this code by reducing it to a single batch. The lack of
shared caching between logged-out and logged-in is not an issue
because we batch the titleinfo queries in WikiModule.php, so whether
the $userBatch cache miss is small or big doesn't make much difference.
It has worked this way for ~10 years and we didn't notice the impact,
so let's simplify the code to recognise how it already worked.

While at it, also fix the missing preload of ext.globalCssJs.user
(User:Admin/global.js). This is trigged by isKnownEmpty in RL\Client
for modules in GROUP_USER, which we did not preload, so there was
always 1 lazy getTitleInfo() call on every logged-in page view.
This now has a chance of being included in the user-specific cache,
as already for ext.globalCssJs.user.styles (User:Admin/global.css).

Bug: T414161
Change-Id: I042daa6c42e8a8d530e0b0c752e3830b98395b18